feature: Job 支持容器执行 - 脚本任务 #2631

+++ src/backend/job-manage/api-job-manage/src/main/java/com/tencent/bk/job/manage/model/web/request/chooser/container/ListContainerByTopologyNodesReq.java
@@ -57,8 +57,8 @@ public class ListContainerByTopologyNodesReq {
     private Map<String, String> podLabels;
 
     @ApiModelProperty(value = "数据起始位置")
-    private Long start;
+    private Integer start;
 
     @ApiModelProperty(value = "拉取数量")
-    private Long pageSize;
+    private Integer pageSize;
 }
